ui changes and custom word wrap
The way I wanted to have the 'cursor' displayed didn't work with the
standard word wrapping because it wont have a space at the end of a
line. So when the cursor is on the space between two words at a wrap
point, it won't be visible.

The text is now manually word wrapped to a specified max column width,
and the moveCursor is hacked together to work with it.

posToListPos converts a position value into a (row,column) value in a
List of stings.